This is chromatography related.
