+++ chemclipse/plugins/org.eclipse.chemclipse.msd.converter.supplier.mzml/src/org/eclipse/chemclipse/msd/converter/supplier/mzml/io/ChromatogramReaderVersion10.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2021, 2024 Lablicate GmbH.
+ * Copyright (c) 2021, 2025 Lablicate GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -271,7 +271,7 @@ public class ChromatogramReaderVersion10 extends AbstractChromatogramReader impl
 				}
 			}
 		}
-		XmlMassSpectrumReader.addTotalSignals(intensities, retentionTimes, chromatogram);
+		XmlChromatogramReader.addTotalSignals(intensities, retentionTimes, chromatogram);
 	}
 
 	private void readContact(MzMLType mzML, IVendorChromatogram chromatogram) {
