+++ chemclipse/plugins/org.eclipse.chemclipse.msd.converter.supplier.mzml/src/org/eclipse/chemclipse/msd/converter/supplier/mzml/io/XmlChromatogramReader.java
@@ -1,0 +1,38 @@
+/*******************************************************************************
+ * Copyright (c) 2021, 2025 Lablicate GmbH.
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ * Matthias Mailänder - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.chemclipse.msd.converter.supplier.mzml.io;
+
+import org.eclipse.chemclipse.msd.converter.supplier.mzml.converter.model.IVendorChromatogram;
+import org.eclipse.chemclipse.msd.converter.supplier.mzml.converter.model.VendorIon;
+import org.eclipse.chemclipse.msd.converter.supplier.mzml.converter.model.VendorScan;
+import org.eclipse.chemclipse.msd.model.core.IIon;
+
+public class XmlChromatogramReader {
+
+	private XmlChromatogramReader() {
+
+	}
+
+	public static void addTotalSignals(double[] intensities, double[] retentionTimes, IVendorChromatogram chromatogram) {
+
+		int tic = Math.min(retentionTimes.length, intensities.length);
+		for(int i = 0; i < tic; i++) {
+			VendorScan scan = new VendorScan();
+			int retentionTime = (int)(retentionTimes[i]);
+			scan.setRetentionTime(retentionTime);
+			float intensity = (float)intensities[i];
+			VendorIon ion = new VendorIon(IIon.TIC_ION, intensity);
+			scan.addIon(ion, false);
+			chromatogram.addScan(scan);
+		}
+	}
+}
